services/custom-deepresearch/requirements.txt
# Norizon search Microservice Dependencies

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.8.0

# HTTP client (async)
httpx>=0.26.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Prompt loading
pyyaml>=6.0

# Logging
structlog>=24.1.0

# Stopwords (for QueryReformulator)
nltk>=3.8.0

# Embeddings (for hybrid search — only needed when ES agent uses hybrid_enabled)
# Uncomment if using Elasticsearch hybrid search:
# sentence-transformers>=2.2.0

# MCP client (for Confluence agent)
mcp>=1.0.0

# Tracing (OpenTelemetry with Phoenix)
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-http>=1.22.0
openinference-semantic-conventions>=0.1.25
